Create a new RelationalProfile for relational non-SQL data stores.
- Moving parts of JdbcTableComponent up into RelationalTableComponent.
- Moving many capabilities from SqlProfile up into RelationalProfile.

An initial version of the (yet to be written) HeapDriver should support
most relational features but it won't understand SQL or any other
text-based query language. Having a useful profile at that level should
allow us to write many TestKit tests in a way that makes them work with
HeapDriver.